remainder of advanced pages
