+++ examples/dotnet/Examples/OpenARealmExamples.cs
@@ -33,7 +33,6 @@ namespace Examples
             config.Schema = new[]
             {
                 typeof(Task),
-                typeof(MyClass),
                 typeof(User)
             };
