+++ source/sdk/dotnet/advanced-guides/bundle-a-realm.txt
@@ -44,7 +44,7 @@ Copy a Realm File
    method to make a copy of the {+realm+} to a new location and/or name. The 
    following code demonstrates this:
 
-   .. literalinclude:: /examples/generated/dotnet/WriteCopy.codeblock.copy_a_realm.cs
+   .. literalinclude:: /examples/generated/dotnet/BundleARealmExamples.codeblock.copy_a_realm.cs
       :language: csharp
 
 
@@ -97,7 +97,7 @@ add code to use it. The code you add depends on the type of app:
       you can do this during start-up of the app; note that it will only run when no 
       {+realm+} file is found at the specified location.
 
-      .. literalinclude:: /examples/generated/dotnet/WriteCopy.codeblock.extract_and_copy_realm.cs
+      .. literalinclude:: /examples/generated/dotnet/BundleARealmExamples.codeblock.extract_and_copy_realm.cs
          :language: csharp
 
    .. tab:: Unity
